historical-graphrag/requirements.txt
openai>=1.0.0
neo4j>=5.0.0
opencc-python-reimplemented>=0.1.7
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.10
numpy>=1.24.0
pyyaml>=6.0
tqdm>=4.65.0
flask>=2.3.0
jieba>=0.42.1
